requirements.txt
numpy==1.15.4
Pillow==5.3.0
opencv-python==3.4.3.18
scikit-learn==0.20.0
pandas==0.23.4
matplotlib
tensorflow
h5py
